Major refactoring and cleanups
- Move bt functions from event handler to bt module
- Move threads from event handler to main
- Start main thread at boot, not bt mode
- Replace printk() with LOG_*() in bt module
- do various other cleanups and adjustments
