+++ hypnos/include/gfx.h
@@ -12,6 +12,7 @@ void gfx_battery_set_label(int);
 void gfx_bt_set_label(int);
 void gfx_time_set_label(char *);
 void gfx_date_set_label(char *);
+void gfx_update(void);
 void lv_task_handler(void);
 
 #endif /* GFX__H */
